Practice Management System/requirements.txt
asgiref==3.8.1
colorama==0.4.6
distlib==0.3.8
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.15.4
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pip==24.2
platformdirs==4.2.2
pluggy==1.5.0
psycopg==3.2.1
psycopg-binary==3.2.1
PyJWT==2.9.0
pytest==8.3.2
pytz==2024.1
PyYAML==6.0.2
setuptools==74.1.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.26.3
